+++ oasislmf/pytools/common/event_stream.py
@@ -254,7 +254,7 @@ class EventReader:
         """
         main_selector = selector_class()
         stream_data = []
-        for file_id, stream_in in enumerate(streams_in):
+        for file_idx, stream_in in enumerate(streams_in):
             mv = memoryview(bytearray(PIPE_CAPACITY))
             byte_mv = np.frombuffer(buffer=mv, dtype='b')
 
@@ -265,7 +265,7 @@ class EventReader:
                     'cursor': 0,
                     'valid_buff': 0,
                     'stream_selector': stream_selector,
-                    'file_id': file_id,
+                    'file_idx': file_idx,
                     }
             stream_data.append(data)
             main_selector.register(stream_in, selectors.EVENT_READ, data)
@@ -304,10 +304,10 @@ class EventReader:
                     byte_mv = data['byte_mv']
                     cursor = data['cursor']
                     valid_buff = data['valid_buff']
-                    file_id = data['file_id']
+                    file_idx = data['file_idx']
                     yield_event = True
                     while yield_event:
-                        cursor, event_id, item_id, yield_event = self.read_buffer(byte_mv, cursor, valid_buff, 0, 0, file_id=file_id)
+                        cursor, event_id, item_id, yield_event = self.read_buffer(byte_mv, cursor, valid_buff, 0, 0, file_idx=file_idx)
 
                         if event_id:
                             self.item_exit()
@@ -317,7 +317,7 @@ class EventReader:
         finally:
             main_selector.close()
 
-    def read_event(self, stream_in, main_selector, stream_selector, mv, byte_mv, cursor, valid_buff, file_id):
+    def read_event(self, stream_in, main_selector, stream_selector, mv, byte_mv, cursor, valid_buff, file_idx):
         """
         read one event from stream_in
         close and remove the stream from main_selector when all is read
@@ -329,7 +329,7 @@ class EventReader:
             byte_mv: numpy byte view of the buffer
             cursor: current cursor of the memory view
             valid_buff: valid data in memory view
-            file_id: file index
+            file_idx: file index
 
         Returns:
             event_id, cursor, valid_buff
@@ -351,7 +351,7 @@ class EventReader:
                         return event_id, cursor, valid_buff
 
                     break
-            cursor, event_id, item_id, yield_event = self.read_buffer(byte_mv, cursor, valid_buff, event_id, item_id, file_id=file_id)
+            cursor, event_id, item_id, yield_event = self.read_buffer(byte_mv, cursor, valid_buff, event_id, item_id, file_idx=file_idx)
 
             if yield_event:
                 if 2 * cursor > valid_buff:
